Project/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
streamlit
pandas
requests
matplotlib
scikit-learn
pytest
presidio-analyzer
spacy
sentence-transformers
